Cleanup annotations, and break up pbar module
