+++ jax_tqdm/__init__.py
@@ -1,1 +1,3 @@
-from jax_tqdm.pbar import PBar, loop_tqdm, scan_tqdm
+from jax_tqdm.base import PBar
+from jax_tqdm.loop_pbar import loop_tqdm
+from jax_tqdm.scan_pbar import scan_tqdm
