--- jax_tqdm/pbar.py
@@ -1,272 +1,0 @@
-from typing import (
-    TYPE_CHECKING,
-    Any,
-    Callable,
-    Generic,
-    Optional,
-    Protocol,
-    TypeVar,
-)
-
-import jax
-import tqdm.auto
-import tqdm.notebook
-import tqdm.std
-from jax.debug import callback
-
-Carry = TypeVar("Carry")
-X = TypeVar("X", int, tuple[int, Any])
-Y = TypeVar("Y")
-R = TypeVar("R")
-
-if TYPE_CHECKING:
-    from dataclasses import dataclass
-else:
-    from chex import dataclass
-
-
-@dataclass
-class PBar(Generic[Carry]):
-    id: int
-    carry: Carry
-
-
-ScanFn = Callable[[Carry, X], tuple[Carry, Y]]
-WrappedScanFn = Callable[[Carry | PBar[Carry], X], tuple[Carry | PBar[Carry], Y]]
-LoopFn = Callable[[int, Carry], Carry]
-WrappedLoopFn = Callable[[int, Carry | PBar[Carry]], Carry | PBar[Carry]]
-
-
-def scan_tqdm(
-    n: int,
-    print_rate: Optional[int] = None,
-    tqdm_type: str = "auto",
-    **kwargs: Any,
-) -> Callable[[ScanFn], WrappedScanFn]:
-    """
-    tqdm progress bar for a JAX scan
-
-    Parameters
-    ----------
-    n : int
-        Number of scan steps/iterations.
-    print_rate : int
-        Optional integer rate at which the progress bar will be updated,
-        by default the print rate will 1/20th of the total number of steps.
-    tqdm_type: str
-        Type of progress-bar, should be one of "auto", "std", or "notebook".
-    **kwargs
-        Extra keyword arguments to pass to tqdm.
-
-    Returns
-    -------
-    typing.Callable:
-        Progress bar wrapping function.
-    """
-
-    update_progress_bar, close_tqdm = build_tqdm(n, print_rate, tqdm_type, **kwargs)
-
-    def _scan_tqdm(func: ScanFn) -> WrappedScanFn:
-        """Decorator that adds a tqdm progress bar to `body_fun` used in `jax.lax.scan`.
-        Note that `body_fun` must either be looping over `jnp.arange(n)`,
-        or be looping over a tuple who's first element is `jnp.arange(n)`
-        This means that `iter_num` is the current iteration number
-        """
-
-        def wrapper_progress_bar(
-            carry: Carry | PBar[Carry], x: X
-        ) -> tuple[Carry | PBar[Carry], Y]:
-            if isinstance(x, tuple):
-                iter_num, *_ = x
-            else:
-                iter_num = x
-
-            if isinstance(carry, PBar):
-                bar_id = carry.id
-                carry_ = carry.carry
-                carry_, x = update_progress_bar((carry_, x), iter_num, bar_id)
-                result = func(carry_, x)
-                result = (PBar(id=bar_id, carry=result[0]), result[1])
-                return close_tqdm(result, iter_num, bar_id)
-            else:
-                carry, x = update_progress_bar((carry, x), iter_num, 0)
-                result = func(carry, x)
-                return close_tqdm(result, iter_num, 0)
-
-        return wrapper_progress_bar
-
-    return _scan_tqdm
-
-
-def loop_tqdm(
-    n: int,
-    print_rate: Optional[int] = None,
-    tqdm_type: str = "auto",
-    **kwargs: Any,
-) -> Callable[[LoopFn], WrappedLoopFn]:
-    """
-    tqdm progress bar for a JAX fori_loop
-
-    Parameters
-    ----------
-    n: int
-        Number of iterations.
-    print_rate: int
-        Optional integer rate at which the progress bar will be updated,
-        by default the print rate will 1/20th of the total number of steps.
-    tqdm_type: str
-        Type of progress-bar, should be one of "auto", "std", or "notebook".
-    **kwargs
-        Extra keyword arguments to pass to tqdm.
-
-    Returns
-    -------
-    typing.Callable:
-        Progress bar wrapping function.
-    """
-
-    update_progress_bar, close_tqdm = build_tqdm(n, print_rate, tqdm_type, **kwargs)
-
-    def _loop_tqdm(func: LoopFn) -> WrappedLoopFn:
-        """
-        Decorator that adds a tqdm progress bar to `body_fun`
-        used in `jax.lax.fori_loop`.
-        """
-
-        def wrapper_progress_bar(
-            i: int, val: Carry | PBar[Carry]
-        ) -> Carry | PBar[Carry]:
-            if isinstance(val, PBar):
-                bar_id = val.id
-                val = val.carry
-                i, val = update_progress_bar((i, val), i, bar_id)
-                result = func(i, val)
-                result = PBar(id=bar_id, carry=result)
-                return close_tqdm(result, i, bar_id)
-            else:
-                i, val = update_progress_bar((i, val), i, 0)
-                result = func(i, val)
-                return close_tqdm(result, i, 0)
-
-        return wrapper_progress_bar
-
-    return _loop_tqdm
-
-
-class UpdateProgressBar(Protocol):
-    def __call__(self, carry: Carry, iter_num: int, bar_id: int) -> Carry:
-        ...
-
-
-class CloseTQDM(Protocol):
-    def __call__(self, result: R, iter_num: int, bar_id: int) -> R:
-        ...
-
-
-def build_tqdm(
-    n: int,
-    print_rate: Optional[int],
-    tqdm_type: str,
-    **kwargs: Any,
-) -> tuple[UpdateProgressBar, CloseTQDM]:
-    """
-    Build the tqdm progress bar on the host
-
-    Parameters
-    ----------
-    n: int
-        Number of updates
-    print_rate: int
-        Optional integer rate at which the progress bar will be updated,
-        If ``None`` the print rate will 1/20th of the total number of steps.
-    tqdm_type: str
-        Type of progress-bar, should be one of "auto", "std", or "notebook".
-    **kwargs
-        Extra keyword arguments to pass to tqdm.
-    """
-
-    if tqdm_type not in ("auto", "std", "notebook"):
-        raise ValueError(
-            'tqdm_type should be one of "auto", "std", or "notebook" '
-            f'but got "{tqdm_type}"'
-        )
-    pbar = getattr(tqdm, tqdm_type).tqdm
-
-    desc = kwargs.pop("desc", f"Running for {n:,} iterations")
-    message = kwargs.pop("message", desc)
-    position_offset = kwargs.pop("position", 0)
-
-    for kwarg in ("total", "mininterval", "maxinterval", "miniters"):
-        kwargs.pop(kwarg, None)
-
-    tqdm_bars = dict()
-
-    if print_rate is None:
-        if n > 20:
-            print_rate = int(n / 20)
-        else:
-            print_rate = 1
-    else:
-        if print_rate < 1:
-            raise ValueError(f"Print rate should be > 0 got {print_rate}")
-        elif print_rate > n:
-            raise ValueError(
-                "Print rate should be less than the "
-                f"number of steps {n}, got {print_rate}"
-            )
-
-    remainder = n % print_rate
-    remainder = remainder if remainder > 0 else print_rate
-
-    def _define_tqdm(bar_id: int) -> None:
-        bar_id = int(bar_id)
-        tqdm_bars[bar_id] = pbar(
-            total=n,
-            position=bar_id + position_offset,
-            desc=message,
-            **kwargs,
-        )
-
-    def _update_tqdm(bar_id: int) -> None:
-        tqdm_bars[int(bar_id)].update(print_rate)
-
-    def _close_tqdm(bar_id: int) -> None:
-        _pbar = tqdm_bars.pop(int(bar_id))
-        _pbar.update(remainder)
-        _pbar.clear()
-        _pbar.close()
-
-    def update_progress_bar(carry: Carry, iter_num: int, bar_id: int) -> Carry:
-        """Updates tqdm from a JAX scan or loop"""
-
-        def _inner_init(_i: int, _carry: Carry) -> Carry:
-            callback(_define_tqdm, bar_id, ordered=True)
-            return _carry
-
-        def _inner_update(i: int, _carry: Carry) -> Carry:
-            _ = jax.lax.cond(
-                i % print_rate == 0,
-                lambda: callback(_update_tqdm, bar_id, ordered=True),
-                lambda: None,
-            )
-            return _carry
-
-        carry = jax.lax.cond(
-            iter_num == 0,
-            _inner_init,
-            _inner_update,
-            iter_num,
-            carry,
-        )
-
-        return carry
-
-    def close_tqdm(result: R, iter_num: int, bar_id: int) -> R:
-        def _inner_close(_result: R) -> R:
-            callback(_close_tqdm, bar_id, ordered=True)
-            return _result
-
-        result = jax.lax.cond(iter_num + 1 == n, _inner_close, lambda r: r, result)
-        return result
-
-    return update_progress_bar, close_tqdm
